[Hetero] Support to set recompute for each stage and micro-batch (#205)
Support to set recompute for each stage and micro-batch.
An example is as follows:
```
recompute_granularity_per_stage_micro_batch:
      - [1, 254, 0, 2, 0]
      - [1, 254, 1, 2, 1]
      - [1, 254, 1, 2, 1]
    recompute_method_per_stage_micro_batch:
      - [1, 254, 0, 2, 0]
      - [1, 254, 0, 2, 0]
      - [1, 254, 1, 2, 1]
    recompute_num_layers_per_stage_micro_batch:
      - [1, 254, 2, 2, 2]
      - [1, 254, 1, 2, 1]
      - [1, 254, 2, 2, 2]
```
The sum of the first item of the sub-lists is equal to
`pipeline_paralell_size`, that is to say, the `1+1+1` is equal to
`pipeline_paralell_size(3)`. The sub-list is organized by the format
`nums-micro-batch, config-flag` except the first item. For this example,
the `254 + 2` is equal to the `num_micro_batch(256)`.
For `recompute_granularity_per_stage_micro_batch`, the `config-flag` can
be:
- `1` for `enable recompute` 
- `0` for `disable recompute`

For `recompute_method_per_stage_micro_batch`, the `config-flag` can be:
- `0` for `uniform` method 
- `1` for `block` method

For `recompute_num_layers_per_stage_micro_batch`, the `config-flag` can
be `0 ~ nums_layers` to set the `recomputed layers`

---------

Co-authored-by: lizhiyu <zyli@baai.ac.cn>