[Tools] Speed up the llavaov data preprocess (#469)
This PR accelerates the data preprocessing of Llava OneVision. It adopts
a multi-task approach, launching multiple tasks simultaneously for
processing, with each task handling a portion of the data. In practical
tests, it can achieve an acceleration of more than twice in the single
image scenario, and an acceleration of up to 1.4 times in the video and
multi-image scenarios.
Usage: set task count by `twice` in  `make_llava_ov_wds.sh`

Co-authored-by: caozhou <caozhou@baai.ac.cn>