[Core] Deprecate legacy hardware adaptation and add readme for the new patching mechanism (#498)
This PR aims to enhance robustness and update the documentation of the
new hardware adaptation mechanism.The following are the specific
changes:

1. Branch Settings have been added for Megatron-LM and Megatron-Energon
2. The "moonlight_deepseek" directory has been added for the ckpt
transformation of deepseek_v3, and users no longer need to pull the
hugging face repo of moonlight
3. The ckpt conversion of llama3 is adapted to the new convert.py, and
"tokenizer_type" has been added in args.py
4. Fixed the issue of excessive loss of first step in qwen2.5-vl
5. The instruction document for the new hardware adaptation mechanism
has been added. The old document requires users to manually checkout to
the specified commit.

---------

Co-authored-by: caozhou <caozhou@baai.ac.cn>