[Tools] Support specified submodule commit for hardware patch (#534)
This PR adds the ability to specify a commit for the submodule and
optionally apply the FlagScale adaptation. Usage is as follows:
`python tools/patch/unpatch.py --backend vllm --backend-commit
<vllm_commit> --no-fs-extension`
`--backend-commit` indicates the commit to which the submodule will
automatically reset after a submodule update. The `--no-fs-extension`
means the FlagScale adaptation will not be applied, and the submodule
will remain in its original state.

Co-authored-by: caozhou <caozhou@baai.ac.cn>