[DistCkpt] Fix the chained optimizer loading error when  resuming from TP/PP to EP (#802)
## Description ##
When using ckpt_format=torch_dist, it is not possible to convert the
TP/PP parallel strategy to EP parallelism.
## Analysis ##
Specifically, when using TP/PP parallelism, the chained optimizer
contains one optimizer. However, when using EP parallelism, it contains
two optimizers: one for non-expert parameters and one for expert
parameters.
The difference in the number of optimizers within the chained optimizer
leads to key indexing errors when loading tp/pp ckpt to ep model.
## How to fix ##
- Modify the sharded_state_dict method of the chained optimizer. When EP
parallelism is enabled yet the checkpoint stems from a TP/PP
configuration, the state_dict of the multiple constituent optimizers are
**merged into a single state_dict** while simultaneously **constructing
a bijective mapping** that records the correspondence between the merged
entries and their original optimizer instances. This consolidation
enables unproblematic checkpoint loading because the chained optimizer
then presents only one optimizer to the restoration routine.
- Modify the load_state_dict method of the chained optimizer so that,
after the unified state_dict has been successfully loaded from the
checkpoint, its parameters are redistributed to the multiple optimizers
currently attached to the model in strict accordance with the
pre-established mapping.

---------

Co-authored-by: Xianduo Li <30922914+lxd-cumt@users.noreply.github.com>